refactor: utils 패키지 내부 클래스 domain으로 이동
