+++ src/main/java/domain/StepGenerator.java
@@ -1,6 +1,4 @@
-package utils;
-
-import domain.StepPoint;
+package domain;
 
 public interface StepGenerator {
