+++ src/test/java/utils/PresentStepGenerator.java
@@ -1,5 +1,6 @@
 package utils;
 
+import domain.StepGenerator;
 import domain.StepPoint;
 
 public class PresentStepGenerator implements StepGenerator {
